Cleaning up and rewriting documentation
This is a first pass over the end-user documentation which will be a published Jekyll site.

The original bootstrap docs will be used as a base, as they provide details on all the components etc. that can be used. The idea is for this documentation to be a one-stop-shop, so rather than redirect people to the components included in Bootstrap by default, and only highlight our custom components, we will show everything in one place, regardless of where it came from.

This means we will simply adapt much of Boostrap's documentation and add our additions as needed. There are some places where more work is required however, the homepage and getting-started page. These have both been significantly modified and i'm more or less happy with them.

The bulk of the work will now be updating the CSS and components pages as needed to integrate Helpful's styles and in clarifying/enhancing the bootstrap documentation wherever this may be lacking.

Although currently this looks like we're just stealing Bootstrap and find/replacing references to it, this is a first pass to tidy up anything we won't be doing (customising downloads for example). As time goes on our own, bespoke, content and customisations will increase and provide actual value over just reading the bootstrap docs themselves.

This does mean we need to careful when bootstrap updates such that we don't have to do all this again to incorporate updates - though this isn't a concern for the short term.

Further documentations will be made in the next few weeks to review everything, at least to a superficial level.