Adding styles from marine portal front-end experiments for xl grid size
These experiments defined two things:

* A new 'XL' grid size
* A 24 rather than 12 column grid

Both of these have been successfully implemented before, but they were tangled up in the front-end experiments application and so not trivial to separate out.

Currently only the new grid size is being implemented, though the code for the grid size is also included (but set to make 12 columns as normal). The the time of this commit the custom grid size doesn't work as required mixins aren't available from bootstrap. This will require loading them from bootstrap as required to fix.