Fixing Less compilation issues
Having imported the right Boostrap variables and mixins the Style Kit Less styles can now be compiled.

Sourcemaps and strict maths are included and compiled css is outputted to both the dist and documentation dist directories.

These tasks need expanding on to include things like autoprefixing, linting etc. This will need to be managed to ensure that the source maps that are generated are not confused by these further transformations.