Finishing documentation for XL grid size
This is a custom grid size designed for large screens. This will mostly be used in web-apps where we know the target audience will have a large screen and we don't want to waste it with useless wide margins.

The bootstrap grid example has been updated to demo this new grid size, when it's available.