Adding compatility support for glyphicons
I don't like doing this as the fonts occupy the root fonts directory (rather than being namespaced) and adds an unnecessary set of additional web-fonts to load.

In the long term it would be nice to remove this dependency without customising bootstrap, but i'm not sure how that could be done.