Adding support for Font-Awesome
This is the first icon-webfont to be included in the BAS Style Kit.

It follows the same loading steps used by typographic fonts (i.e. Less file imported in BAS Style Kit styles and font's copied into namespaced directory).

There will likely be a number of additional, more specialised, icon fonts to be included as well. This may cause a resource size issue in future as the amount of data/resources we pull in increases - particularly when south or on mobile. Therefore we may want to talk an alternative approach whereby all icon-fonts (and possibly typographic fonts as well) are optional rather than included out of the box.

To display the available icons within the documentation a gulp task is used to produce a yaml file of just the class names. This is then looped through to produce a grid of icons. This is the same process used by Bootstrap but written in gulp.