Adding map glyphs icon web-font
We also use this web-font on the homepage to replace the snowflake with an outline of antarctica.

Note, this font is licensed and is not distributed publicly, it is therefore included within this project as a 'native' font. This isn't ideal, but unless we get in touch with the author there is little way to resolve this.