Adding BAS navbar
First integration of Helpful's styles with bootstrap.

This required more than just copying and pasting - as far as possible the bootstrap inverse navbar was used as a base for creating two additional variations, a 'default' navbar using the BAS styles and a 'large' version using larger fonts.

This required deconstructing Helpful's styles, changing the selectors used for some things and replacing static values with either static or calculated variables. In some cases Helpful's styles weren't copied because they were specific to the public website only or would require making substantial changes to bootstrap for very little results. These alterations are documented and will need to be discussed.

An example of using this nabber has been added to the docs and as a standalone example. More examples will need updating in future to show off the different positioning options.