Adding styles for footer component
This is not yet complete - namely this needs documenting. The footer component is based on a navbar but supports general content without the need for explicit classes for things like links and text.

Based on the helpful template example but using font icons for social media links and official brand colours. Examples are in the starter template currently until this is cleaned up.