Updating to bootstrap 3.3.5
This has required no style changes (though updated bootstrap styles will now be used).

The documentation hasn't needed to be updated as we were already usign the 3.3.5 docs (but the 3.3.4 styles as this is controlled by NPM).

This 'upgrade' therefore cannot be considered a test of upgrading to a new version of Boostrap.