Switching to 24 column grid
This is a change carried over from the Marine Portal Frontend Experiments bootstrap customisations. 12 columns didn't give enough control, and whilst 24 is possibly too many, it makes it simple to 'scale up'.

Changing this within the Less/CSS is a simple variable change, however this will then require all documentation and examples to be updated - which is more resource intensive.

This is the last major change required for the first 0.x release. So far the documentation and Less styles have been updated. Only the grid example has been updated to ensure the new grid works. All other examples will need to converted as part of the general example updates.