Adding versioning support for production documentation
In production environments multiple versions can be accessed through version sub-directories (e.g. 0.1.0/index.html and 2.3.0/index.html etc.).

This requires the latest version and the environment (production/development/etc.) to be identified. For future compatbility this information is kept in a separate jekyll data file, rather than in _config. This allows ansible to template out this file in the future if needed.

Currently this information is set manually (with documentation updated to explain this process) as the documentation can only be built in development environments. Unless ansible was used to switch environmens (which would be annoying when developing) it is easiest to make this manual as it is infrequent and can be added to an existing release workflow.

Documentation can now be environment aware and shows different banners and a version selector if in production. A version index is created and moved to become the site index in such environments. Both the version selector and index are populated from a common list of versions, stored as a Jekyll data file.

Because of the site structure changes, links can no longer use "../" safely, they must refer to the current version. This does not apply within documentation, unless they refer to the homepage (for that version) as "/" as this is now the version index.