Partially resolving colour scheme issues
This at least gets things into a sensible place for a review to take place on how to resolve the issues around colour schemes.

This also includes a Reference Image produced by Pete which helps clarify, visually, the differences between the various colours that are defined and which should be actually included in the 'resolved' and final colour scheme.