Rethinking CI/CD setup
I think I fundermentally misunderstood how testing and deployment environments are related in Semaphore - in that there is no relationship, other than the same git commit being checked out.

This means any build commands to generate artefacts (such as a static website) need to repeated in the deployment environment as needed.

This seems odd because this is not the typical way in which testing/deployment is used in larger projects. In our case our tests are sort of embedded within the deployment process, and aren't run independently. This will change as we add site checking (relative links etc.) and more JS testing (as this is added to future versions of the Style Kit).

Currently I haven't split out the CSSLint task into an independent task, but this should be done in the near future if this new approach works.

To simplify changing what happens in deployment and in testing I have extracted everything possible to Ansible playbooks so this can be better versioned and changed more easily in future.