Setting build environment for use in CI/CD
This is limited currently to flipping between staging and development (but these are the only environments that use CI/CD currently).

In future we would need to allow setting a production environment, possibly by using the current branch as an environment variable, or using a simple shell script to set one.

This requires updating all checks made against the 'site.data.provisioned.env' variable to be updated.