Adding build revision to staging environment
This uses an environment variable set by Semaphore to represent the revision being built. This is to aid to diagnosing if 'stale' revisions are being deployed.

The plugin added to expose an environment variable as a Jekyll variable is general purpose, and so could be used to define other information outside of Jekyll, but display within the site, if needed.