Adding a pretend-master branch check to simulate master deploys
To test things that will run against the Master branch we need a 'pretend' branch so we don't need to use the actual Master branch.

This is actually fairly easy now we use a script to determine which environment we are in. Ideally this will still be solved by per-branch environment variables supported in Semaphore.