Adding continous integration instructions for the master branch
These are the same as those used for develop, but with the addition of a 'pretend-master' branch.

This lets us test changes that will apply to the master branch, without actually comitting them to master (and misleading people into thinking it is a release). This may not be needed once things are further developed, but for now this is needed.

Better documentation is needed on setting up and using this branch.