Adding initial updated deployment process for production
This isn't finished yet (particuarly in terms of documentation) but will essentially mirror the deployment process used for staging to production.

This won't be enough however as production has the extra complication of needing to handle multiple versions of the documentation. This will be added later as well as CDN deployment.