Updating documentation to meet new versioning system
Versions will now exist in a versions directory, with an index populated by directories within this. Older versions will need to be copied into this directory for them to appear. They will not be stored as generated sites within the repository.

This ensures the root of the site will always be the current version of the documentation.