Fixing generation of versions index
The Jekyll plugin took its list of versions from the site directory, which of course doesn't exist until you build the site.

We now check the versions directory in the site source, which isn't as clean as we need to track at least pointers to each version (but not their contents) in source control but is better than it not working.

In the longer term we may need to think of a better way to do this.