Adding new IAM users and policies for deployments
This removes the need to use a single IAM user with permissions to deploy to both Production and Staging environments.

This is in preparation to moving to using Semaphore's S3 deployment strategy, at which point the legacy IAM user and policy can be removed. This means Ansible Vault is then no longer needed for this project.

Our Terraform configuration is also tidied up.