Switching to Semaphore's S3 deployment method
This means we no longer need to store S3 credentials (and so can remove Ansible Vault) and cleans up our playbooks.

The CDN deployment is a bit harder as we need to use an Ansible variable to ensure the assets are uploaded to the right place. This may be something that needs to be revisited in the future.