Resetting project on Pristine base project
This commit represents the beginning of the process to base the Style Kit on the project template, Pristine.

Many of the files removed in this commit will be restored in future commits as they are re-integrated.

This commit also marks the beginning of the 0.2.0 alpha. The major change being switching from Less to Sass. This process will take place in stages and likely require some elements of the style kit to be re-implemented.

The aim initially is to produce a version of the style kit compatible with the current 0.1.0 release but using Sass. After which new changes scheduled as part of 0.2.0 can be added.

The motivation to switch to Sass stems from Bootstrap's decision to use Sass as their primary preprocessor starting with 4.0. Although this is not yet available it seems sensible to convert the style kit now so integrating with version 4 is less painful than also having to switch css language.