Porting Less styles to Sass
This required porting all the customisations made to bootstrap (now bootstrap-sass) and unique styles included in the Style Kit.

In most cases this only meant using the variable syntax and adopting the conventional directory structure for Sass projects. For some aspects, such as grids, which use complex mixins, more extensive changes were needed.

There are several limitations with this port, and as such is not yet complete.
1) Fontface related styles have not been ported, as these require copying font files etc. this will happen as a second step.
2) Other than ensuring the Sass files compile no testing of this port has been completed. It is expected the majority of styles will transfer without issue (since it is not complicated code), however areas such as grids will require extensive testing to ensure compatibility is not broken with existing CSS.

Gulp task runner tasks required to compile the custom Bootstrap css and the Style Kit css are reintroduced, along with associated requirements (node).