Adding local development server for testing
As the main documentation for the Style Kit will be external to this project, we still need a quick way to test changes made to styles do what they should.

Rather than build a set of minimal documentation it is only really necessary to have a set of minimal HTML files served by a very basic server.

To that end a set of minimal examples (which will be improved upon and include all the examples from the current documentation site) will be included in this project, along with a minimal web server using node.