Fixing variable precedence issue
Style Kit variables need to override those defined in bootstrap to create our custom bootstrap version.

In Less its fine to load the variables after they are used within files (Less effectively leaves the variables as placeholders until the end of the compile).

In Sass the variables must be loaded before the files they apply to (Sass effectively fixes the value of variables as it processes each file).

Both these approaches are logical, once you know which is needed.