Adding staging environment
This includes using Terraform to create a deployment user in AWS IAM for Semaphore to deploy into the development environment of the BAS CDN.

It also includes generally refactored CI/CD workflows to use the simplified nature of this project (i.e. it only needs to use Gulp now) so can leverage the Semaphore provided instances of Node.