Adding Docker support
Implements #1.

Uses Docker compose to define two containers - the first contains the build environment for the Style Kit (Gulp and its plugins), the other exposes the testbed content, including the files built by the first container.

The ‘bas-style-kit’ image built for this project, is kept in the BAS Docker Registry (provided by the BAS GitLab instance).

This removes the need for Ansible, as this was only used for the local development environment.

In time Continuous Integration and Deployment (CI/CD) will be moved to the BAS GitLab instance, and will replace Semaphore.

Some Gulp tasks have had to be revised to support Docker (specifically mounted volumes can’t be removed, just their content) - and new tasks added to fit the Docker workflow (using watch to detect file changes etc.).

The ‘gulp-sass’ Gulp plugin had to be updated to use a more recent version of the Sass compiler, as older versions didn’t natively support Alpine, and so would need to built a module from source, which would have required a significant number of built tools (including Python!).

There will be rough edges surrounding this work, and so these changes should be viewed as a first step - rather than a polished feature.