Re-enabling App container and versioning image
The app container was disabled to test the web container, and should been been un-commented before.

The project image (app image) is now versioned using the date of the last build, to ensure users aren’t using a ‘latest’ version, which could mean anything and lead to problems when used by multiple developers.