Updating Node Modules
The only out of date dependency is our StyleLint configuration, which has changed fairly considerably.

Testing this new version locally raised no new warnings however, and resolves the deprecation warning for a rule which will be removed in future StyleLint versions - which ensures we are now ready for this change.

As Node Modules have changed we need to rebuild the project image.