Adding ground work for testbed
This uses the Nunjucks templating system (Node implementation of Jinja2).

Gulp tasks are used to compile templates, which are then exposed through the Nginx image. Tasks are also used to gather a list of samples and collections (sets of samples).

These samples and collections are kept as isolated ‘snippets’ rendered into full pages (in ‘/samples’) and collections (in ‘/collections’).

In this commit, only two samples are defined to prove this setup will work.