Improving testbed architecture
Macros are expanded to include options for setting links to isolated examples (for use in lists), or the site index (for use in isolated examples).

A flag is now set within lists to inform a sample if it is rendered within a list, or standalone. This is used to set options with the section/section-header macro, but could be used for other purposes as well in the future.

Notes section removed from lists as it isn’t obvious what this would be used for, i.e. what notes would we put in a list that we wouldn’t list with the individual sample?