Preparing for GitLab CI review apps
This requires the ability to host the testbed at an arbitrary path (as review apps will be subdirectories per review).

It also requires a new provisioning environment to provide a place to host review apps.

New CI jobs have been added to prepare the tested, upload it to a S3 bucket for each review, and to delete it again later.