Fixing using brand images and text together
New fix class added to use both together. Incompatibility was caused by ‘display: block’ for images.

Using Git blame to find why this was set gave no real explaination [1]. Removing this option didn’t seem to cause any regressions in terms of images loosing their height etc. when used on there own.

As a wider task once support for IE9 is removed, it would therefore be a good idea to more throughly test if this property can simply be overriden in our styles, and remove the need for a fix class.

[1] https://github.com/twbs/bootstrap/commit/8e2348e9eda51296eb680192379ab37f10355ca3