Switching from develop/master to master/tagged-releases
To better fit with GitLabs Git model the old ‘develop/master’ branches are removed in favour of a single ‘master’ branch (representing mostly what ‘develop’ was, with the exception that it should build), and tags off of this representing merges against the old ‘master’ branch.

This simplifies things and means we no longer have massive merge requests against master during a release (this upcoming release will be the last one).

Features and bugs are still developed in their own branches, and merged in - but there is no longer an intermediate branch.

If needed, release branches can be long term if two versions of the Style Kit need to be maintained for a short period for example.

This requires changes to the repository, which will be made alongside this commit, and to the CI/CD process (captured in .gitlab-ci.yml).