Merging Terraform resources into a single environment
This is part of a general trend across all projects to manage all Terraform resources in a single set of configuration files, rather than split into different environments.

The Style Kit this largely means moving IAM inline policies into a single place, as a common user was already defined. With the exception of the Testbed this project also relies purely on other projects for its storage and distribution so has a more limited set of resources to manage.