Moving SRI task to processing stage
GitLab runs jobs within a stage together so the assets to be processed don’t exist if this job is run during the build phase.

It wouldn’t make sense to include this in the ‘lint’ stage, so a new ‘process’ stage is added for general tasks like this.

It would have been ideal to include in the ‘package’ stage, but we want this file to be included in the package made.