Adding NPM support
Previously this was an ad-hoc process and is now at least documented, though not automated.

As NPM versions cannot be changed once published it is important to manually check things are in order, and is therefore still a manual process.

A NPM ignore file is included to remove testbed resources from the package.

As the NPM package is packed within a Docker container, additional meta files need to be included in the Docker image so they are available.