Updating Style Kit variables
Since we have integrated the Bootstrap and Style Kit styles together, we need to update hte variables we use to set things like rounded edges. In the past these variables were all prefixed, so won’t override the default Bootstrap variables.

This is no longer needed, and so these prefixes are removed (except for colours and some other specific variables). The Style Kit styles used to apply these prefixed-variables are no longer needed as a result of this.