Overriding responsive utilities for large breakpoint
We need to change the Bootstrap style directly due to how CSS styles are applied.

The rule for the large breakpoint needs to be changed to add a ‘max-width’ rule, otherwise it clobbers the ‘XL’ breakpoint added within the Style Kit.

This override was not previously needed, so further work is needed to look at not needing to do this in the future, as it is not elegent.