Correcting using a border colour for primary buttons
In the Style Kit documentation primary buttons had a distinct border colour. The Style Kit itself currently doesn’t make a distinction, which creates an inconsistency.

Since we documented there was a difference, and this is present in the Bootstrap default styles (for their primary button) this different border has been implemented in the Style Kit.

This technically represents a breaking change - but shouldn’t have been something users relied on (since it was never supposed to the way it was).

This difference can be of course be removed, but that should be reflected in the documentation as well.