Redeveloping image purchase/copyright component
This component has been completed redeveloped  based on styles from BBC News. It removes all of the inconsistencies with different media breakpoints and produces a much cleaner implementation.

This component has also been renamed, to reflect its more general purpose. Given the new legal policies being adopted by BAS it is necessary to identify which resources (if any) are not covered by the Open Government License. This component is now intended to do that (i.e. showing an identifier that can then be found in the accompanying copyright policy). It can still be used for image purchasing (via a linked attribution) if applicable.