Adding samples for GDS colours
Currently these new colour scheme are temporary as they need to be integrated with the main Style Kit.

A decision needs to be made on which red we use for example. Some of these will be quite simple, others will require more thought, particularly in terms of what uses a colour.