Adding testbed support for experimental samples
This is designed to differentiate samples which are temporary in nature used to research a problem, or test candidates for a new feature.

Such samples will be kept in case they are useful for future use, but won’t be kept up to date with changes to the rest of the Style Kit.