Adding javascript for cookie notice
Consists of one native script, to handle the cookie notice itself (i.e. by recording the notice is acknowledged) and a copy of the Bootstrap alert.js with class names updated (i.e. ‘.alert’ to ‘.bsk-alert’).

Similar override scripts will be needed for other interactivity, and these will be added gradually.

This component represents the first JS files included in the Style Kit and this will have a number of knock-on effects, some of which are addressed now, but not all. Any required Gulp tasks to expose javascript files have been added for example, but others for minifying and linting these will be added later.