Correcting experimental button hover colour
This was actually caused by how we load bootstrap styles in our button group styles, this had the effect of overriding our styles.

We now load our styles again (only button styles are affected) to override the incorrect override - which is inelegant but reduces the chances of side-effects in the future and may be addressed with wider efforts to remove duplicate styles in the Style Kit.