Reclassifying bootstrap and fonts as dependencies
I.e. Not dev-dependencies.

The logic being that if you are using this package as part of another, you will be compiling the styles yourself and there’s no ‘dist’ provided which would integrate things like Open Sans etc.

Therefore you will need these packages available to use the styles, but you don’t need to use our Gulp setup for compiling them (you could use web-pack or something).

This means we need to distinguish between packages *we* use for *building* the dist files, vs. those used as inputs to that process, which are ‘dependencies’.

This difference is clearer if we listed things like ‘jQuery’ as a dependency, since we depend on it, but don’t use it as part of our build process.