Gulp is no longer a global module
This is a best practice to keep projects more self-contained and avoid conflicts where multiple projects might need the same module at different versions.

Having switched to Yarn, it is more forceful about not using global packages and so it’s probably best we follow this advice.

The effect of this change is a longer path needed when calling gulp commands, which since we rarely run gulp commands manually, isn’t a significant problem.

Note: CI will need updating, but as our Gulp tasks are all going to be rewritten there isn’t much point doing that now.