Updating stylelint configuration
THe Stylelint standard configuration changed as part of updating to the latest version and introduced rule options which our styles failed to meet.

Most can be resolved as a false positive (incorrectly disallowing ‘@include’ for example), some others have required code changes to a small number of files, but none of these changes are considered major or breaking.