Re-adding archive task
This generates a zip archive of the dist directory ussed as part of a release.

By generating this file with a generic name and renaming it within CD, we no longer need to zip the files within CI using an OS package.

This means we can drop the dependency on the deprecates ‘bas-aws’ package (in favour of the GDS image).