Beginning to replace the Testbed
This is a work in progress and involves replacing the testbed, and its samples, with a new, separated, environment.

So far this consists of a set of gulp tasks which can generate the testbed assets (css, including imports from the main style kit) and a prototype function for extracting information out of samples for building up a dynamic index of samples.

The samples themsleves will need to rewritten (as the format will change and nunjucks is deprecated). This will be done in future commits once the new format is established.

As Nginx was only used for the testbed, this has been moved to the Testbed.