Re-adding all testbed gulp functionality
This adds gulp tasks to compile templates (using Pug rather than Nunjucks) and create an index of samples and sample collections.

Samples now use pug frontmatter to express settings such as the sample title (the number is loaded from the sample name).

Collections are also expressed within each sample, rather than as a separate set of files.

Overall these new functions are significantly simplier and more elegant that the previous versions.

An index of samples is compiled during the main sample build task, rather than as a standalone task.

A layout and includes need to be made before real samples can start to be re-added (and converted to Pug).