Adding Pug layouts for a standard style kit template
This includes base layouts (both generic and Style Kit specific) and more opinionated layouts for a typical page layout with a header/footer using a fixed width design.

Includes are used for specific elements such as the cookie notice, etc.

This has several limitations, including harded code values for things such as the site name, and a lack of options in terms of fixed/fluid containers.