Adding variables for template layouts/includes
This is part of preparing for a standalone set of Style Kit templates using Pug.

This means allowing the brand text in the navbar to be customised for example.

Some variables such as the CSS URL aren’t expected to be changed, but since these are variables with default values, this flexibility can be ignored if not required.

However this does not add everything needed for a set of templates as anything not needed for the testbed has been skipped (such as navbar items).

Variables exist at different levels (as the base and standard layouts suit different aims) and can be overwritten as needed in ‘lower level’ templates.