Adding a namespace to all URLs
One of the issues with old testbed was the lack of a consistent path between local builds (which sat on the root URL) and hosted versions for commits or branches (which used the commit/branch as a sub-folder).

This made logic difficult to use between the two kinds of environment, and ultimately meant links in hosted versions typically didn’t work.

To overcome this in the new testbed, all environments will use a namespace, with local environments using a static placeholder namespace ‘develop’.

This ensures that the structure of the site remains the same between all environments, allowing links to be tested locally using a namespace, and removing the need for different handling logic.