Switching to OGL licensing
Under the terms of the MIT license (used by bootstrap) we can incorporate their work into our own, providing we acknoledge their licensing statement (including copyright).

We retain the rights to our changes on top of this, which we can license as we wish, in this case the Open Government License (OGL).