Setting drop-down active colour to match navbar variant
I.e. grey instead of blue.

This means we now have two active variant colours, which is not ideal.

We need to look in more detail about whether we want to create a monochrome design (and extend this to things like tabs/pills more widely), or that colour is ok.

We could have a half-way house so only the navbar (which is dominated by it’s dark colours) uses a dark active colour, since otherwise this would clash - but keep regular drop-downs colourful.

Personally - I would keep things monochrome as we already do for things like default buttons. Colour should catch the eye, which is harder if it’s also part of a base design.