Updating dependencies
Updates packages to new versions (and CDN references in testbed samples) and switches to Yarn 2.

Yarn 2 changes how packages are stored, removing the traditional node_modules directory. This breaks imports we used to make to files within this directory (notably for fonts, bootstrap and the Pug theme in the testbed).

Until a more elegant solution can be found, these dependencies have been manually copied into this project as with the exception of the Pug theme, they are likely to change in the future.